+++ internal/mods/system/dal/role.dal.go
@@ -62,7 +62,7 @@ func (m roleRepo) Update(ctx context.Context, role *dto.RolePB, options ...dto.R
 	return dto.ConvertRole2PB(saved), nil
 }
 
-func (m roleRepo) List(ctx context.Context, in *pb.ListRolesRequest, options ...dto.RoleQueryOption) ([]*dto.RolePB, int64, error) {
+func (m roleRepo) List(ctx context.Context, in *pb.ListRolesRequest, options ...dto.RoleQueryOption) ([]*dto.RolePB, int32, error) {
 	var option dto.RoleQueryOption
 	if len(options) > 0 {
 		option = options[0]
@@ -92,13 +92,13 @@ func NewRoleDal(db *Data, logger log.Logger) dto.RoleRepo {
 	}
 }
 
-func rolePageQuery(ctx context.Context, query *ent.RoleQuery, in *pb.ListRolesRequest, option dto.RoleQueryOption) ([]*dto.RolePB, int64, error) {
+func rolePageQuery(ctx context.Context, query *ent.RoleQuery, in *pb.ListRolesRequest, option dto.RoleQueryOption) ([]*dto.RolePB, int32, error) {
 	if in.OnlyCount {
 		count, err := query.Count(ctx)
 		if err != nil {
 			return nil, 0, err
 		}
-		return nil, int64(count), nil
+		return nil, int32(count), nil
 	}
 
 	query = roleQueryPage(query, in)
@@ -108,7 +108,7 @@ func rolePageQuery(ctx context.Context, query *ent.RoleQuery, in *pb.ListRolesRe
 		return nil, 0, err
 	}
 	result, err := query.All(ctx)
-	return dto.ConvertRoles(result), int64(count), err
+	return dto.ConvertRoles(result), int32(count), err
 }
 
 func roleQueryPage(query *ent.RoleQuery, in *pb.ListRolesRequest) *ent.RoleQuery {
