+++ internal/mods/system/service/user.gin.go
@@ -53,7 +53,7 @@ func (m UserAPIGINService) ListUsers(context *gin.Context, request *pb.ListUsers
 	var res gins.Result
 	if err == nil {
 		res.Success = true
-		res.Total = int64(response.TotalSize)
+		res.Total = response.TotalSize
 		res.Data = response.Users
 	}
 	resp.Result(context, res, err)
